+++ lib/features/settings/components/rows/premium_introduction.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
-import 'package:pilll/features/premium_introduction/ab_test/a/premium_introduction_sheet2.dart';
+import 'package:pilll/features/premium_introduction/ab_test/a/premium_introduction_sheet.dart';
 import 'package:pilll/utils/analytics.dart';
 import 'package:pilll/components/atoms/font.dart';
